+++ src/test/java/yym/svydovets/algorithm/task/csosvita/search_loopy/OneTenHundredTest.java
@@ -4,12 +4,12 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
-class InfiniteSequenceOfDigitsTest {
+class OneTenHundredTest {
 
     @Test
     void findNumbers() {
         int[] arr = new int[]{3,14,7,6};
-        final String result = InfiniteSequenceOfDigits.findNumbers(arr);
+        final String result = OneTenHundred.findNumbers(arr);
         assertEquals("0 0 1 0", result);
     }
 }
